@@ -10,11 +10,14 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=10"
+		"node": ">=12"
 	},
 	"scripts": {
-		"test": "xo && nyc ava && tsd"
+		"//test": "xo && nyc ava && tsd",
+		"test": "xo && ava && tsd"
 	},
 	"files": [
 		"index.js",
@@ -34,10 +37,10 @@
 		"buffer"
 	],
 	"devDependencies": {
-		"ava": "^2.4.0",
+		"ava": "^3.15.0",
 		"nyc": "^15.1.0",
-		"tsd": "^0.13.1",
-		"xo": "^0.35.0"
+		"tsd": "^0.14.0",
+		"xo": "^0.37.1"
 	},
 	"nyc": {
 		"reporter": [
